FusionScan/requirements.txt
Flask~=3.1.0
Flask-Login~=0.6.3
Flask-SQLAlchemy~=3.1.1
Flask-Migrate~=4.1.0
Werkzeug~=3.1.3
opencv-python
dlib==19.24.1
face-recognition==1.3.0
numpy==1.26.4
pandas~=2.2.3
xlsxwriter~=3.2.0
pyserial~=3.5